1st round of code review fixes
