+++ .github/workflows/go.yml
@@ -34,4 +34,4 @@ jobs:
       run: go build -v .
 
     - name: Test
-      run: go test -v github.com/rjar2020/post-delivery/tests
+      run: go test -v github.com/rjar2020/post-delivery/service
